Fixing like product functionality

+++ client/angular-app/src/app/modules/product/product-details/product-details.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { map } from 'rxjs/operators';
 
 import { ProductService } from '../product.service';
@@ -18,6 +18,7 @@ export class ProductDetailsComponent implements OnInit {
 
   constructor(
     private activatedRoute: ActivatedRoute,
+    private router: Router,
     private productService: ProductService,
     private userService: UserService
   ) { }
@@ -26,8 +27,12 @@ export class ProductDetailsComponent implements OnInit {
     return this.userService.isCurrentLoggedUserOwnerOfProduct(this.user.username, this.userService.getCurrentUserName());
   }
 
-  get isCurrentLoggedUserAdmin():boolean {
+  get isCurrentLoggedUserAdmin(): boolean {
     return this.userService.getCurrentUserName() === 'ADMIN';
+  }
+
+  get isCurrentUserHasAlreadyLikedTheProduct(): boolean {
+    return this.product.peopleLikedProduct.includes(this.userService.getCurrentUserName()) ? true : false;
   }
 
   ngOnInit(): void {
@@ -45,12 +50,27 @@ export class ProductDetailsComponent implements OnInit {
   }
 
   productDeleteHandler(): void {
-    this.productService.deleteProduct(this.product._id)
-      .subscribe(
-        // ЗА НОТИФИКАЦИЯ ЩЕ МИ ТРЯБВА RESPONSA ОТ БАЗАТА
-        error => console.log(error),
-        () => console.log('Stream has been closed!')
-      );
+    if (window.confirm("Are you sure that you want to delete this product?")) {
+      this.productService.deleteProduct(this.product._id)
+        .subscribe(
+          // ЗА НОТИФИКАЦИЯ ЩЕ МИ ТРЯБВА RESPONSA ОТ БАЗАТА
+          response => this.router.navigate(['/home']),
+          error => console.log(error),
+          () => console.log('Stream has been closed!')
+        );
+    }
+  }
+
+  productBuyHandler(): void {
+    if (window.confirm("Are you sure that you want to buy this product?")) {
+      this.productService.deleteProduct(this.product._id)
+        .subscribe(
+          // ЗА НОТИФИКАЦИЯ ЩЕ МИ ТРЯБВА RESPONSA ОТ БАЗАТА
+          response => this.router.navigate(['/home']),
+          error => console.log(error),
+          () => console.log('Stream has been closed!')
+        );
+    }
   }
 
   likeProductHandler(): void {
